Remove the hack which does not fix blurred images in Chrome anymore
The problem appears only in Chrome and is caused by Chrome's
inability to downscale images without blurring them.
Since version 92 the hack used (transform: transaleZ(0)) does not work,
so I made the images slightly bigger with 'transform: scale()'
but this breaks the page on bigger touchscreen devices as the images
become too big.