Fix contact form submission
- Add logging when PHP Mailer fails to sent message
due to error different than exception.

- Rememove PHP Mailer setFrom() which was erroneously
used to pass email address of the person using the
contact form. This would make PHP Mailer throw an error
(different from exception), as setFrom() must be the email
of the owner of the email account used to send the contact
form message.

- Reformat message subject to include sender's name and
email address.

- Amend custom error message when exception is thrown.

- Refactor.